Tie in Elm frontend
All of the frontend generation is located in the frontend/ dir. To
generate new static files (placed into frontend/dist/ and symlinked for
Rocket to find in static/) run `npm run build`. No actual functionality
is present besides loading the correct files.

This should be a great starting point to begin actually building the web
application.