Add error handling into request handlers
Now both login and register routes accept JSON and return JSON to better
integrate with Elm. In addition, the models module was renamed to model
to better fit the singular current standard for the rest of the modules.
(small style difference I know). Error handling has been added in
another module named error, and from now on no more unwrap, expect, or
other functions calls that could panic are allowed. There are still a
few (I think) in the code right now, but those NEED to be removed asap.

Serde was added to work with JSON through Rocket, and it doesn't make
sense to have both rustc-serialize and serde (they both do
serialization). The rustc-serialize dep hasn't been removed, but should
be soon.

Errors for the login and register function do not currently reflect what
the actual errors are that should be returned to the user.

Errors should probably be returned through a response that is JSON, not
plain text, to stay consistent with other responses to the client.