Check for confirmation on /dash route
This commit also updates a few crates, notably jsonwebtoken to version
2. This allows us to move web token validation into that crate, as it is
now supported by the crate itself. Also, it allows us to remove the
dependency on rustc_serialize, as jwt uses serde now. This fixes #15.

The confirmation part fixes #16.

The entire request guard for auth was rewritten to reduce the number of
routes that are needed for protected routes. This allowed the entire
un-authed route for GET /dash to be removed and refactored into a single
route.