Create tests on Travis CI for most code
Dependencies were updated as well.

Rocket and co. are now being run from the git repo, which should be changed ASAP. This was to prevent dependency issues with serde, but it means that the API is unstable and not documented well. Lots of weird issues with cookies popped up after this change, but I could not find corresponding commits in Rocket that would explain them.